Move filter criterion handlers into separate file
